+++ R/pptx_slide_manip.R
@@ -1,18 +1,18 @@
+
 #' @export
 #' @title Add a slide
 #' @description Add a slide into a pptx presentation.
-#' @param x an rpptx object
-#' @param layout slide layout name to use
-#' @param master master layout name where \code{layout} is located
-#' @examples
-#' my_pres <- read_pptx()
-#' layout_summary(my_pres)
-#' my_pres <- add_slide(my_pres,
-#'   layout = "Two Content", master = "Office Theme"
-#' )
-#' @seealso [print.rpptx()], [read_pptx()], [plot_layout_properties()], [ph_with()], [layout_summary()]
-#' @family functions slide manipulation
-add_slide <- function(x, layout = "Title and Content", master = "Office Theme") {
+#' @param x an `rpptx` object.
+#' @param layout slide layout name to use.
+#' @param master master layout name where `layout` is located.
+#' @param ... Key-value pairs of the form `"short form location" = object` passed to [phs_with]. See section
+#'   `"Short forms"` in [phs_with] for details, available short forms and examples.
+#' @param .dots List of key-value pairs of the form `list("short form location" = object)`. Alternative to `...`. See
+#'   [phs_with] for details.
+#' @example inst/examples/example_add_slide.R
+#' @seealso [print.rpptx()], [read_pptx()], [layout_summary()], [plot_layout_properties()], [ph_with()], [phs_with()]
+#' @family slide_manipulation
+add_slide <- function(x, layout = "Title and Content", master = "Office Theme", ..., .dots = NULL) {
   slide_info <- x$slideLayouts$get_metadata()
   slide_info <- slide_info[slide_info$name == layout & slide_info$master_name == master, ]
 
@@ -24,6 +24,25 @@ add_slide <- function(x, layout = "Title and Content", master = "Office Theme") 
       ". Layout names should not be duplicated."
     )
   }
+
+  dots_list <- list(...)
+  if (length(dots_list) > 0 && !is_named(dots_list)) {
+    cli::cli_abort(
+      c("Missing key in {.arg ...}",
+        "x" = "{.arg ...} requires a key-value syntax, with a ph short-form location as the key",
+        "i" = "Example: {.code add_slide(x, ..., title = 'My title', 'body[1]' = 'My body')}"
+      )
+    )
+  }
+  if (length(.dots) > 0 && !is_named(.dots)) {
+    cli::cli_abort(
+      c("Missing names in {.arg .dots}",
+        "x" = "{.arg .dots} must be a named list, with ph short-form locations as names",
+        "i" = "Example: {.code add_slide(x, ..., .dots = list(title = 'My title', 'body[1]' = 'My body'))}"
+      )
+    )
+  }
+
   new_slidename <- x$slide$get_new_slidename()
 
   xml_file <- file.path(x$package_dir, "ppt/slides", new_slidename)
@@ -35,8 +54,13 @@ add_slide <- function(x, layout = "Title and Content", master = "Office Theme") 
   x$content_type$add_slide(partname = file.path("/ppt/slides", new_slidename))
 
   x$slide$add_slide(xml_file, x$slideLayouts$get_xfrm_data())
-
   x$cursor <- x$slide$length()
+
+  # fill placeholders
+  dots <- utils::modifyList(dots_list, .dots %||% list())
+  if (length(dots) > 0) {
+    x <- phs_with(x, .dots = dots)
+  }
   x
 }
 
@@ -64,7 +88,7 @@ add_slide <- function(x, layout = "Title and Content", master = "Office Theme") 
 #'
 #' file <- tempfile(fileext = ".pptx")
 #' print(doc, target = file)
-#' @family functions slide manipulation
+#' @family slide_manipulation
 #' @seealso [read_pptx()], [ph_with()]
 on_slide <- function(x, index) {
   l_ <- length(x)
@@ -79,9 +103,9 @@ on_slide <- function(x, index) {
   location <- which(x$slide$get_metadata()$name %in% filename)
 
   x$cursor <- x$slide$slide_index(filename)
-
   x
 }
+
 
 #' @export
 #' @title Remove a slide
@@ -95,7 +119,7 @@ on_slide <- function(x, index) {
 #' my_pres <- read_pptx()
 #' my_pres <- add_slide(my_pres)
 #' my_pres <- remove_slide(my_pres)
-#' @family functions slide manipulation
+#' @family slide_manipulation
 #' @seealso [read_pptx()], [ph_with()], [ph_remove()]
 remove_slide <- function(x, index = NULL, rm_images = FALSE) {
   l_ <- length(x)
@@ -144,7 +168,7 @@ remove_slide <- function(x, index = NULL, rm_images = FALSE) {
 #' x <- add_slide(x)
 #' x <- ph_with(x, "Hello world 2", location = ph_location_type())
 #' x <- move_slide(x, index = 1, to = 2)
-#' @family functions slide manipulation
+#' @family slide_manipulation
 #' @seealso [read_pptx()]
 move_slide <- function(x, index = NULL, to) {
   x$presentation$slide_data()
@@ -253,5 +277,86 @@ ensure_slide_index_exists <- function(x, slide_idx) {
         "x" = "Presentation has {cli::no(n)} slide{?s}."
       ), call = NULL
     )
+  }
+}
+
+
+# internal workhorse get/set slide visibility
+# x : rpptx object
+# slide_idx: id of slide
+# value: Use TRUE / FALSE to set visibility.
+.slide_visible <- function(x, slide_idx, value = NULL) {
+  stop_if_not_rpptx(x)
+  slide <- x$slide$get_slide(slide_idx)
+  slide_xml <- slide$get()
+  node <- xml2::xml_find_first(slide_xml, "/p:sld")
+  if (is.null(value)) {
+    value <- xml2::xml_attr(node, "show")
+    value <- as.logical(as.numeric(value))
+    ifelse(is.na(value), TRUE, value) # if show is not set, the slide is shown
+  } else {
+    stop_if_not_class(value, "logical", arg = "value")
+    xml2::xml_set_attr(node, "show", value = as.numeric(value))
+    slide$save()
+    invisible(x)
+  }
+}
+
+
+#' Get or set slide visibility
+#'
+#' PPTX slides can be visible or hidden. This function gets or sets the visibility of slides.
+#' @param x An `rpptx` object.
+#' @param value Boolean vector with slide visibilities.
+#' @rdname slide-visible
+#' @export
+#' @example inst/examples/example_slide_visible.R
+#' @return Boolean vector with slide visibilities or `rpptx` object if changes are made to the object.
+`slide_visible<-` <- function(x, value) {
+  stop_if_not_rpptx(x)
+  stop_if_not_class(value, "logical", arg = "value")
+  n_vals <- length(value)
+  n_slides <- length(x)
+  if (n_vals > n_slides) {
+    cli::cli_abort("More values ({.val {n_vals}}) than slides ({.val {n_slides}})")
+  }
+  if (n_vals != 1 && n_vals != n_slides) {
+    cli::cli_warn("Value is not length 1 or same length as number of slides ({.val {n_slides}}). Recycling values.")
+  }
+  value <- rep(value, length.out = n_slides)
+  for (i in seq_along(value)) {
+    .slide_visible(x, i, value[i])
+  }
+  invisible(x)
+}
+
+
+#' @param hide,show Indexes of slides to hide or show.
+#' @rdname slide-visible
+#' @export
+slide_visible <- function(x, hide = NULL, show = NULL) {
+  stop_if_not_rpptx(x)
+  idx_in_both <- intersect(as.integer(hide), as.integer(show))
+  if (length(idx_in_both) > 1) {
+    cli::cli_abort(
+      "Overlap between indexes in {.arg hide} and {.arg show}: {.val {idx_in_both}}",
+      "x" = "Indexes must be mutually exclusive.")
+  }
+  if (!is.null(hide)) {
+    stop_if_not_integerish(hide, "hide")
+    stop_if_not_in_slide_range(x, hide, arg = "hide")
+    slide_visible(x)[hide] <- FALSE
+  }
+  if (!is.null(show)) {
+    stop_if_not_integerish(show, "show")
+    stop_if_not_in_slide_range(x, show, arg = "show")
+    slide_visible(x)[show] <- TRUE
+  }
+  n_slides <- length(x)
+  res <- vapply(seq_len(n_slides), function(idx) .slide_visible(x, idx), logical(1))
+  if (is.null(hide) && is.null(show)) {
+    res
+  } else {
+    x
   }
 }